backend/requirements.txt
flask
flask-cors
pymongo[srv]
gunicorn
python-dotenv
certifi
